[refactor] #2 dto 위치 변경
