+++ src/main/java/com/abcdedu_backend/member/dto/request/LoginRequest.java
@@ -1,17 +1,15 @@
-package com.abcdedu_backend.member.controller.dto.request;
+package com.abcdedu_backend.member.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
-public record SignUpRequest(
-        @NotBlank
-        String name,
+public record LoginRequest(
         @Email
         @NotBlank
         String email,
         @Size(min = 6, max = 20)
         @NotBlank
         String password
-){
+) {
 }
