+++ src/main/java/com/abcdedu_backend/member/dto/response/ReissueResponse.java
@@ -1,4 +1,4 @@
-package com.abcdedu_backend.member.controller.dto.response;
+package com.abcdedu_backend.member.dto.response;
 
 public record ReissueResponse(
         String accessToken
